requirements.txt
tensorflow==2.15.0
keras==2.15.0
huggingface_hub==0.20.0
datasets==2.16.0
transformers==4.36.0
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
Pillow==10.1.0
opencv-python==4.8.1.78
pyyaml==6.0.1